♻️ [Refactoring]: conversion-optionsテストをテストルールに基づいて分割
テストルール（Kent C. Doddsのベストプラクティス）に従い、
conversion-options.test.tsを機能ごとに7つのファイルに分割：

- conversion-options.default.test.ts: デフォルト値のテスト
- conversion-options.merge.test.ts: マージ機能のテスト
- conversion-options.validate.test.ts: バリデーション機能のテスト
- conversion-options.normalize.test.ts: 正規化機能のテスト
- conversion-options.typeguards.test.ts: 型ガード関数のテスト
- conversion-options.colormode.test.ts: カラーモード関連のテスト
- conversion-options.from.test.ts: fromファクトリー関数のテスト

フラットな構造を採用し、可読性と保守性を向上
